+++ index.js
@@ -2,12 +2,20 @@ const express = require("express");
 const mongoose = require("mongoose");
 const cookieSession = require("cookie-session");
 const passsport = require("passport");
+const bodyParser = require("body-parser");
+var cors = require("cors");
 const keys = require("./config/keys");
 require("./model/User");
 require("./services/passport");
 
 mongoose.connect(keys.mongoURI, { useNewUrlParser: true });
 const app = express();
+
+//handles incoming data
+app.use(bodyParser.json());
+
+//handles cors
+// app.use(cors());
 
 //this code will make the cookie session for a given time with the given keys
 app.use(
@@ -25,6 +33,19 @@ app.use(passsport.session());
 // authRoutes(app);
 //the below line is the same than the above 2 lines
 require("./routes/authRoutes")(app);
+require("./routes/billingRoutes")(app);
+
+if (process.env.NODE_ENV === "production") {
+  //express will serve up production assets
+  // like main.js or main.css
+  app.use(express.static("client/build"));
+
+  //Express will serve up the index.html
+  //if it does not recognize the route like /surveys
+  app.get("*", (req, res) => {
+    res.sendFile(path.resolve(__dirname, "client", "build", "index.html"));
+  });
+}
 
 //setting up the PORT to be used
 const PORT = process.env.PORT || 5000;